Add files via upload
This is my main code for now, it allows me to generate sets of J values which can be saved and then later read. I can then use these sets to have consistent initial conditions for anything I want to produce with the data. i.e. a calibration curve or for later comparison to simulated annealing.